pyproject.toml
[project]
name = "mobile-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Balakrishnan"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "flask-pymongo (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
